stock-client/src/main/java/cz/tut/client/stockclient/WebClientStockClient.java
package cz.tut.client.stockclient;

import cz.tut.client.stockclient.model.StockPrice;
import lombok.extern.slf4j.Slf4j;
import org.springframework.web.reactive.function.client.WebClient;
import reactor.core.publisher.Flux;

import java.io.IOException;
import java.time.Duration;

@Slf4j
public class WebClientStockClient {

    private final WebClient webClient;

    public WebClientStockClient(WebClient webClient) {
        this.webClient = webClient;
    }

    public Flux<StockPrice> pricesFor(String symbol) {
       return  webClient.get()
                .uri("http://localhost:8099/stocks/{symbol}",symbol)
                .retrieve()
                .bodyToFlux(StockPrice.class)
                .retryBackoff(5, Duration.ofSeconds(1),Duration.ofSeconds(20))
                .doOnError(IOException.class, e -> log.error(e.getMessage()));
    }
}
